mcp-servers/lol-mcp-server/requirements.txt
fast-mcp
requests
python-dotenv 